refactor: applied sourcery suggestions
