+++ src/korone/modules/translator/handlers/utils/errors.py
@@ -1,0 +1,10 @@
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright (c) 2024 Hitalo M. <https://github.com/HitaloM>
+
+
+class TranslationError(Exception):
+    pass
+
+
+class QuotaExceededError(TranslationError):
+    pass
